+++ src/main/java/br/com/example/kazuhiro/controletransaction/modules/user/service/UserBalanceServiceInterface.java
@@ -5,5 +5,5 @@ import java.util.UUID;
 import br.com.example.kazuhiro.controletransaction.modules.user.entitites.UserEntity;
 
 public interface UserBalanceServiceInterface {
-  UserEntity applyTransaction(UUID userId, String tipo, int valor);
+  UserEntity applyTransaction(UUID userId, String tipo, Long valor);
 }
